Make Chart.yaml pass linting
`appVersion` should be a string, but was read as a float64. The docs recommend to use quotes, see https://helm.sh/docs/topics/charts/.

`icon` is recommended. Is set to the openstad.org favicon for now. The URL isn't pretty due to it being an uploaded attachment in Apostrophe. So if we can change this to openstad.org/icon.svg that would be even better.