Pass MongoDB connection string env var to frontend & auth
These can be added separately for the frontend & auth service through the values file under `secrets.mongodb.frontendConnectionString` and `secrets.mongodb.authConnectionString`. It is allowed to use `{database}` in the connection string, which will be replaced in the frontend service at runtime — to allow multiple sites to be hosted on one frontend service — and in the auth service at configuration time (replaced with the `sessions` database by default).